--- app/src/main/java/com/crypto/application/RateQuery.java
@@ -1,4 +1,0 @@
-package com.crypto.application;
-
-public class RateQuery {
-}
