--- app/src/main/java/com/crypto/infrastructure/adapters/outbound/http/BinanceApiDto.java
@@ -1,29 +1,0 @@
-package com.crypto.infrastructure.adapters.outbound.http;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-
-public class BinanceApiDto {
-
-
-    @JsonProperty("s")
-    private String symbol;
-    @JsonProperty("p")
-    private double price;
-
-    public BinanceApiDto(String symbol, double price) {
-        this.symbol = symbol;
-        this.price = price;
-    }
-
-    BinanceApiDto() {
-    }
-
-    public String getSymbol() {
-        return symbol;
-    }
-
-    public double getPrice() {
-        return price;
-    }
-}
