--- app/src/test/java/com/crypto/domain/service/CryptoServiceTest.java
@@ -1,6 +1,0 @@
-package com.crypto.domain.service;
-
-// TODO write test
-class CryptoServiceTest {
-
-}
