+++ src/main/java/com/crypto/domain/model/Rate.java
