+++ src/main/java/com/crypto/infrastructure/adapters/inbound/dto/RateDto.java
