+++ src/main/java/com/crypto/infrastructure/adapters/outbound/repository/CryptoRepository.java
