+++ src/main/java/com/crypto/infrastructure/adapters/outbound/repository/RateRepository.java
