+++ src/main/java/com/crypto/infrastructure/adapters/outbound/repository/entity/CryptoEntity.java
