+++ src/test/resources/__files.json/bnbusdt-response.json
@@ -1,0 +1,13 @@
+{
+  "e": "trade",
+  "E": 1697533856921,
+  "s": "BNBUSDT",
+  "t": 674947545,
+  "p": "212.80000000",
+  "q": "0.30600000",
+  "b": 4841092401,
+  "a": 4841092826,
+  "T": 1697533856920,
+  "m": true,
+  "M": true
+}
