+++ src/test/resources/__files.json/btcusdt-response.json
@@ -1,0 +1,13 @@
+{
+  "e": "trade",
+  "E": 1697533859420,
+  "s": "BTCUSDT",
+  "t": 3242655744,
+  "p": "28402.75000000",
+  "q": "0.00021000",
+  "b": 22773752789,
+  "a": 22773751323,
+  "T": 1697533859420,
+  "m": false,
+  "M": true
+}
