+++ src/test/resources/__files.json/ethusdt-response.json
@@ -1,0 +1,13 @@
+{
+  "e": "trade",
+  "E": 1697533859967,
+  "s": "ETHUSDT",
+  "t": 1211107434,
+  "p": "1587.92000000",
+  "q": "0.03170000",
+  "b": 14425188775,
+  "a": 14425189674,
+  "T": 1697533859967,
+  "m": true,
+  "M": true
+}
