+++ src/test/resources/__files.json/eurusdt-response.json
@@ -1,0 +1,13 @@
+{
+  "e": "trade",
+  "E": 1697533859511,
+  "s": "EURUSDT",
+  "t": 102311082,
+  "p": "1.05290000",
+  "q": "23.30000000",
+  "b": 303857236,
+  "a": 303857119,
+  "T": 1697533859511,
+  "m": false,
+  "M": true
+}
