+++ src/test/resources/__files.json/xrpusdt-response.json
@@ -1,0 +1,13 @@
+{
+  "e": "trade",
+  "E": 1697533859776,
+  "s": "XRPUSDT",
+  "t": 557200221,
+  "p": "0.49110000",
+  "q": "40.00000000",
+  "b": 5568281689,
+  "a": 5568280716,
+  "T": 1697533859775,
+  "m": false,
+  "M": true
+}
