+++ src/test/resources/application-test.properties
@@ -1,0 +1,2 @@
+spring.jpa.show-sql=true
+api_host=http://localhost:9090
