--- web-client/build.gradle
@@ -1,38 +1,0 @@
-plugins {
-    id 'java'
-    id 'org.springframework.boot' version '3.1.3'
-    id 'io.spring.dependency-management' version '1.1.3'
-}
-
-group 'com.crypto'
-version '0.0.1-SNAPSHOT'
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    implementation 'org.springframework.data:spring-data-jpa:3.1.4'
-    implementation 'org.springframework.boot:spring-boot-starter-data-jpa:3.0.4'
-    implementation 'org.springframework.boot:spring-boot-starter-web:3.1.3'
-    implementation 'org.json:json:20230618'
-    //implementation 'org.springframework.boot:spring-boot-starter-websocket:3.1.4'
-
-    implementation 'org.java-websocket:Java-WebSocket:1.5.4'
-
-
-    implementation 'org.springframework:spring-beans:6.0.12'
-    implementation 'org.springframework:spring-web:6.0.12'
-    implementation 'com.fasterxml.jackson.core:jackson-annotations:2.15.2'
-    implementation 'org.springframework.boot:spring-boot-starter-web'
-    //implementation 'org.projectlombok:lombok:1.18.22'
-
-    runtimeOnly 'com.h2database:h2'
-
-    testImplementation 'org.junit.jupiter:junit-jupiter-api:5.8.1'
-    testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.8.1'
-}
-
-test {
-    useJUnitPlatform()
-}
